Fixed reset board function (again)
There were two bugs:
1. When I reset the board, the first player was often
'O', not 'X'.
2. Sometimes, I wasn't able to fill in a square with
a value.

The first issue was fixed by using the length of the
game array to determine the next player. If the gameArray
length (after removing undefined elements) is an even
number, then the next player is 'X'. If the gameArray
length (after removing undefined elements) is an odd
number, then the next player is 'O'.

The second issue was fixed by using $('#0').html('')
instead of ('#0').empty()